requirements.txt
bcrypt>=3.2.0
Django>=3.1.5
django-cors-headers>=3.6.0
django-rest-framework>=0.1.0
PyJWT>=2.0.0
python-dotenv>=0.15.0
boto3>=1.16.63
Pillow>=8.1.0
celery>=5.0.5
redis>=3.5.3